Use all cores - make that scalable
This is a rather important change, but that was planned for a while now

-Implement eris-sharder to take care of the clustering and sharding 
-Add index.js as the master process
-Add main.js as the main process (obviously lul)
-Transfer the ready event to main.js, as it is emitted 
before the launch call, and thus wouldn't be triggered 
the first time
-Most importantly, the automatic setup of the database for the first run has been deleted and i don't plan on bringing it back. It will need a manual setup of the database through the dashboard of rethinkDB before the first run.

**Tests passed**
-The overall process works, tested launching and running the ping command 
**Tests needed**
Most of the stuff has not been tested yet, for example launching without the database to see how the process arguments are handled, or using commands/features that requires the database and access to the data of all clients.

There is still a lot to change, especially logging wise, and IPC wise. There's gonna be a need to review the caching of the database, atm all the clusters would cache the database, resulting in an excessive amount of unnecessary RAM used