Inter-cluster reload, redact as helper and help
The reload command of the previous commits had a serious issue: As each clusters are a separated process, the reload only affected one cluster, so other clusters were still using outdated files. 

To prevent that, a lot of changes have been done:
-Moved the reload methods to the Reloader helper class
-Added the broadcastReload() method to the IPCHandler and made it handle the two events "reload" and "reloadDone"

The reload  command now doesn't actually reload, it just parse arguments and call for broadcastReload(), then each clusters call the according method of the Reloader helper class to reload the specified file. Once done, the clusters report back with a boolean if the reload was successful or not (if not, they log themselves the error)

Along with this, a few other changes were introduced:

-Finished the help command, it now can output the specified command help
-Added the embedColor option in the config, and reflected this change on all commands outputting embeds
-Added the redact helper function (moved from repl)
-Added the getRandomNumber module
-Added the eval command on the way

**Tests passed**
-The eval command work as expected
-As it can be seen in the index, by forcing the sharding manager to have only one guild per shard and spawing 2 clusters, i managed to have 2 guilds on 2 different clusters, and could confirm that the reload command works as expected
-The help command works as expected, regardless of the parameters