Remove unnecessary conversions
When first extending the users database entries,  the idea of "rolling-back" the extension came to give to rethinkdb a proper object, ensuring it wouldn't try to save the entire class.

For that purpose, the simple idea of removing the methods and stringifying it to JSON was first implemented, although, for some reasons, rethinkdb couldn't process it, so i just hardcoded a fix, which was to parse it after stringifying it.

(Note: JSON.stringify() is supposed to handle functions when stringifying an object and remove them, it didn't work in this case and somehow made the call-stack size to exceed its limits, which is why it is manually done)

Later, when extending the guildEntry, i made it cleaner so it would return a normal object without the methods, and give that to rethinkDB, as i forgot to do the same for extended user entries, this commit fix that